requirements.txt
django
Pillow
numpy
scikit-image
google-cloud-translate
google-cloud-vision
djangorestframework
django-cors-headers
sklearn
